moved navbar to folder
