+++ python-project/react-app/public/index.html
@@ -3,7 +3,7 @@
   <head>
     <meta charset="utf-8" />
     <link rel="icon" type="image/x-png" href="favicon.png"/>
-    <title>Simple React App</title>
+    <title>Insta-Clone</title>
   </head>
   <body>
     <div id="root"></div>
